Create const file for cortexm
